src/simulators/requirements.txt
# kafka-python has a bug with Python 3.12
# kafka-python-ng is the maintained fork that fixes it
kafka-python-ng==2.2.2
